+++ lib/src/model/filter/filter_column_item.dart
@@ -21,19 +21,12 @@ class FilterColumnsItem extends FilterItem {
     columns[column] = shown;
   }
 
-  @override
-  List<Map<String, dynamic>> applyFilter(List<Map<String, dynamic>> data) {
+  List<Map<String, dynamic>> removeColumns(List<Map<String, dynamic>> data) {
     columns.forEach((column, show) {
       if (!show) {
         data.forEach((item) => item.remove(column.$name));
       }
     });
     return data;
-  }
-
-  @override
-  String toString() {
-    final amountOfColumnsEnabled = columns.values.where((show) => show).length;
-    return 'FilterColumnsItem: ($isEnabled), amountOfColumns: $amountOfColumnsEnabled';
   }
 }