+++ lib/src/screen/moor_table_viewer_widget.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' as flutter;
 import 'package:moor_db_viewer/src/model/filter/filter_data.dart';
 import 'package:moor_db_viewer/src/navigator/db_navigator.dart';
+import 'package:moor_db_viewer/src/style/theme_dimens.dart';
 import 'package:moor_db_viewer/src/viewmodel/global_viewmodel.dart';
 import 'package:moor_db_viewer/src/viewmodel/moor_table_viewer_viewmodel.dart';
 import 'package:moor_db_viewer/src/widget/provider/provider_widget.dart';
@@ -19,8 +21,7 @@ class MoorTableViewerWidget extends StatefulWidget {
   _MoorTableViewerWidgetState createState() => _MoorTableViewerWidgetState();
 }
 
-class _MoorTableViewerWidgetState extends State<MoorTableViewerWidget>
-    implements MoorTableViewerNavigator {
+class _MoorTableViewerWidgetState extends State<MoorTableViewerWidget> implements MoorTableViewerNavigator {
   final _key = GlobalKey();
 
   @override
@@ -29,37 +30,56 @@ class _MoorTableViewerWidgetState extends State<MoorTableViewerWidget>
       consumer: (context, viewModel, child) => Scaffold(
         key: _key,
         appBar: AppBar(
+          centerTitle: Theme.of(context).platform == TargetPlatform.iOS,
           title: Text(viewModel.title),
           actions: [
             IconButton(
-              icon: Icon(Icons.filter_list),
-              onPressed: viewModel.onFilterClicked,
-            )
+              icon: Icon(Icons.refresh),
+              onPressed: viewModel.onRefreshClicked,
+            ),
+            Stack(
+              children: [
+                Center(
+                  child: IconButton(
+                    icon: Icon(Icons.filter_list),
+                    onPressed: viewModel.onFilterClicked,
+                  ),
+                ),
+                Positioned(
+                  top: ThemeDimens.padding16,
+                  right: ThemeDimens.padding12,
+                  child: Container(
+                    height: ThemeDimens.padding8,
+                    width: ThemeDimens.padding8,
+                    decoration: BoxDecoration(
+                      shape: BoxShape.circle,
+                      color: viewModel.hasFilter ? Theme.of(context).accentColor : Colors.transparent,
+                    ),
+                  ),
+                )
+              ],
+            ),
           ],
         ),
         body: LayoutBuilder(
           builder: (context, constraints) {
-            if (!viewModel.hasColumns && viewModel.hasData)
-              return Center(
-                  child: Text(
-                      'No columns selected for the`${viewModel.tableName}` table'));
-            if (!viewModel.hasData)
-              return Center(
-                  child: Text(
-                      'No data added to the `${viewModel.tableName}` table'));
+            if (viewModel.error != null) return Center(child: Text(viewModel.error));
+            if (!viewModel.hasColumns && viewModel.hasData) return Center(child: Text('No columns selected for the`${viewModel.tableName}` table'));
+            if (!viewModel.hasData && viewModel.hasFilter) return Center(child: Text('No data found for your current filter on the `${viewModel.tableName}` table'));
+            if (!viewModel.hasData) return Center(child: Text('No data added to the `${viewModel.tableName}` table'));
             return SingleChildScrollView(
               scrollDirection: Axis.vertical,
               child: SingleChildScrollView(
                 scrollDirection: Axis.horizontal,
                 child: DataTable(
                   columns: [
-                    for (final item in viewModel.filteredData[0].keys)
+                    for (final item in viewModel.data[0].keys)
                       DataColumn(
                         label: Text(item),
                       ),
                   ],
                   rows: [
-                    ...viewModel.filteredData.map((item) => DataRow(
+                    ...viewModel.data.map((item) => DataRow(
                           cells: [
                             ...item.keys.map(
                               (key) => DataCell(
@@ -75,18 +95,14 @@ class _MoorTableViewerWidgetState extends State<MoorTableViewerWidget>
           },
         ),
       ),
-      create: () => MoorTableViewerViewModel()
-        ..init(this, Provider.of<GlobalViewModel>(context).db, widget.table),
+      create: () => MoorTableViewerViewModel()..init(this, Provider.of<GlobalViewModel>(context).db, widget.table),
     );
   }
 
   @override
-  Future<void> goToFilter(
-      TableInfo<moor.Table, DataClass> table, FilterData filteredData) async {
-    final newFilterData = await DbViewerNavigator.of(context)
-        .goToTableFilter(table, filteredData);
+  Future<void> goToFilter(TableInfo<moor.Table, DataClass> table, FilterData filteredData) async {
+    final newFilterData = await DbViewerNavigator.of(context).goToTableFilter(table, filteredData);
     if (newFilterData == null) return;
-    Provider.of<MoorTableViewerViewModel>(_key.currentContext, listen: false)
-        .updateFilter(newFilterData);
+    Provider.of<MoorTableViewerViewModel>(_key.currentContext, listen: false).updateFilter(newFilterData);
   }
 }