+++ lib/src/viewmodel/moor_table_viewer_viewmodel.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:moor_db_viewer/src/model/filter/filter_data.dart';
-import 'package:moor_db_viewer/src/model/filter/filter_limit_results_item.dart';
 import 'package:moor_flutter/moor_flutter.dart';
 import 'package:moor_flutter/moor_flutter.dart' as moor;
 
@@ -10,24 +9,25 @@ class MoorTableViewerViewModel with ChangeNotifier {
   TableInfo<moor.Table, DataClass> _table;
 
   final _data = List<Map<String, dynamic>>();
-  final _filteredDataList = List<Map<String, dynamic>>();
   var _filteredData = FilterData();
 
   int totalResults = 0;
 
-  List<Map<String, dynamic>> get filteredData => _filteredDataList;
+  String error;
 
-  bool get hasData => _filteredDataList.isNotEmpty;
+  bool get hasFilter => _filteredData.hasFilters;
 
-  bool get hasColumns => hasData && _filteredDataList[0].keys.isNotEmpty;
+  List<Map<String, dynamic>> get data => _data;
 
-  String get title =>
-      '${_table.entityName} ($totalResults-${_filteredDataList.length})';
+  bool get hasData => _data.isNotEmpty;
+
+  bool get hasColumns => hasData && _data[0].keys.isNotEmpty;
+
+  String get title => '${_table.entityName} ($totalResults-${_data.length})';
 
   String get tableName => _table.entityName;
 
-  Future<void> init(MoorTableViewerNavigator navigator, GeneratedDatabase db,
-      TableInfo<moor.Table, DataClass> table) async {
+  Future<void> init(MoorTableViewerNavigator navigator, GeneratedDatabase db, TableInfo<moor.Table, DataClass> table) async {
     _navigator = navigator;
     _db = db;
     _table = table;
@@ -35,36 +35,36 @@ class MoorTableViewerViewModel with ChangeNotifier {
   }
 
   Future<void> _getData() async {
-    // todo find a better way to acces the database for no this is fine
-    // ignore: invalid_use_of_protected_member, invalid_use_of_visible_for_testing_member
-    final result1 = await _db.customSelectQuery(
-        'SELECT COUNT(*) FROM ${_table.actualTableName}',
-        readsFrom: {_table}).get();
-    totalResults = result1.first.data['COUNT(*)'];
-    // todo find a better way to acces the database for no this is fine
-    // ignore: invalid_use_of_protected_member, invalid_use_of_visible_for_testing_member
-    final query = _db.select(_table);
-    final limitFilter = _filteredData.filters[0];
-    if (limitFilter is FilterLimitResultsItem) {
-      query.limit(limitFilter.limit);
+    try {
+      error = null;
+      notifyListeners();
+      // todo find a better way to acces the database for no this is fine
+      // ignore: invalid_use_of_protected_member, invalid_use_of_visible_for_testing_member
+      final result1 = await _db.customSelectQuery('SELECT COUNT(*) FROM ${_table.actualTableName}', readsFrom: {_table}).get();
+      totalResults = result1.first.data['COUNT(*)'];
+      // todo find a better way to acces the database for no this is fine
+
+      final sqlQuery ='SELECT * FROM ${_table.actualTableName} ${_filteredData.getWhere()} ${_filteredData.getLimit()}';
+      print(sqlQuery);
+      final result = await _db.customSelectQuery(sqlQuery, readsFrom: {_table}).get();
+      final _newData = result.map((item) => item.data).toList();
+      final _correctDisplayData = _filteredData.removeColumns(_newData);
+      _data
+        ..clear()
+        ..addAll(_correctDisplayData);
+    } catch (e) {
+      error = e.toString();
+    } finally {
+      notifyListeners();
     }
-    final result = await query.get();
-    final _newData = result.map((item) => item.toJson()).toList();
-    _data
-      ..clear()
-      ..addAll(_newData);
-    final newFilterData = _filteredData?.applyFilter(_newData) ?? _newData;
-    _filteredDataList
-      ..clear()
-      ..addAll(newFilterData);
-    notifyListeners();
   }
+
+  void onRefreshClicked() => _getData();
 
   void onFilterClicked() => _navigator.goToFilter(_table, _filteredData);
 
   void updateFilter(FilterData filterData) {
     _data.clear();
-    _filteredDataList.clear();
     _filteredData = filterData;
     notifyListeners();
     _getData();
@@ -72,6 +72,5 @@ class MoorTableViewerViewModel with ChangeNotifier {
 }
 
 abstract class MoorTableViewerNavigator {
-  void goToFilter(
-      TableInfo<moor.Table, DataClass> table, FilterData _filteredData);
+  void goToFilter(TableInfo<moor.Table, DataClass> table, FilterData _filteredData);
 }