+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-## [0.0.2] - 04-04-2020
+## [0.0.1] - 04-04-2020
 #Added:
+- A refresh button on the table viewer
 - Gif to indicate what the packages does.
 - Added readme instructions
 
 #Fixed: 
+- Filtering is now done at database level
 - Fixed a bug where the wrong screen was used as entry point
 - Renamed the example folder
 - Updated the example to use only allowed methods